Removed parent_institution
The Document model has been updated. The database is designed as follows: An Institution holds ReferenceNumbers which hold Documents. Before this edit, the document saved both the RefNumber parent and the institution which is already set by the reference number.
Removed this field and updated files accordingly.
This is still work in progress.